grpc-service-sample refactored. It compiles
