+++ .gitignore
@@ -3,12 +3,12 @@ graphql-api-sample/target/
 rest-api-sample/target/
 soap-api-sample/target/
 grpc-interface/target/
-grpc-service/target/
+grpc-service-sample/target/
 !/graphql-api-sample/.mvn/wrapper/maven-wrapper.jar
 !/rest-api-sample/.mvn/wrapper/maven-wrapper.jar
 !/soap-api-sample/.mvn/wrapper/maven-wrapper.jar
 !/grpc-interface/.mvn/wrapper/maven-wrapper.jar
-!/grpc-service/.mvn/wrapper/maven-wrapper.jar
+!/grpc-service-sample/.mvn/wrapper/maven-wrapper.jar
 !/graphql-api-sample/**/src/main/**/target/
 !/rest-api-sample/**/src/main/**/target/
 !/soap-api-sample/**/src/main/**/target/
@@ -16,49 +16,49 @@ grpc-service/target/
 !/rest-api-sample/**/src/test/**/target/
 !/soap-api-sample/**/src/test/**/target/
 !/grpc-interface/**/src/test/**/target/
-!/grpc-service/**/src/test/**/target/
+!/grpc-service-sample/**/src/test/**/target/
 
 ### STS ###
 graphql-api-sample/.apt_generated
 rest-api-sample/.apt_generated
 soap-api-sample/.apt_generated
 grpc-interface/.apt_generated
-grpc-service/.apt_generated
+grpc-service-sample/.apt_generated
 graphql-api-sample/.classpath
 rest-api-sample/.classpath
 soap-api-sample/.classpath
 grpc-interface/.classpath
-grpc-service/.classpath
+grpc-service-sample/.classpath
 graphql-api-sample/.factorypath
 rest-api-sample/.factorypath
 soap-api-sample/.factorypath
 grpc-interface/.factorypath
-grpc-service/.factorypath
+grpc-service-sample/.factorypath
 graphql-api-sample/.project
 rest-api-sample/.project
 soap-api-sample/.project
 grpc-interface/.project
-grpc-service/.project
+grpc-service-sample/.project
 graphql-api-sample/.settings
 rest-api-sample/.settings
 soap-api-sample/.settings
 grpc-interface/.settings
-grpc-service/.settings
+grpc-service-sample/.settings
 graphql-api-sample/.springBeans
 rest-api-sample/.springBeans
 soap-api-sample/.springBeans
 grpc-interface/.springBeans
-grpc-service/.springBeans
+grpc-service-sample/.springBeans
 graphql-api-sample/.sts4-cache
 rest-api-sample/.sts4-cache
 soap-api-sample/.sts4-cache
 grpc-interface/.sts4-cache
-grpc-service/.sts4-cache
+grpc-service-sample/.sts4-cache
 
 ### VS Code ###
 graphql-api-sample/.vscode/
 rest-api-sample/.vscode/
 soap-api-sample/.vscode/
 grpc-interface/.vscode/
-grpc-service/.vscode/
+grpc-service-sample/.vscode/
 .vscode/
