+++ grpc-service-sample/pom.xml
@@ -2,20 +2,24 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.6.8</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.faxterol</groupId>
-	<artifactId>grpcdemo</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>grpcdemo</name>
-	<description>Demo project for Spring Boot</description>
+
+	<groupId>com.nmatute</groupId>
+	<artifactId>grpc-service-sample</artifactId>
+	<version>1.0.0</version>
+	<name>grpc-service-sample</name>
+	<description>gRPC Service Sample</description>
+
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 	</properties>
+
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -43,19 +47,19 @@
 			<scope>test</scope>
 		</dependency>
 
-		<!-- gRPC Demo -->
+		<!-- gRPC dependencies -->
 		<dependency>
 			<groupId>io.github.lognet</groupId>
 			<artifactId>grpc-spring-boot-starter</artifactId>
 			<version>4.7.1</version>
 		</dependency>
+
 		<dependency>
-			<groupId>com.faxterol.grpcdemo</groupId>
-			<artifactId>grpc-interfaces</artifactId>
-			<version>1.0-SNAPSHOT</version>
+			<groupId>com.nmatute</groupId>
+			<artifactId>grpc-interface</artifactId>
+			<version>1.0.0</version>
 		</dependency>
 
-		<!-- Si usas Java 11, este paquete no viene de forma nativa, incluyelo en tu pom file -->
 		<dependency>
 			<groupId>javax.annotation</groupId>
 			<artifactId>javax.annotation-api</artifactId>
