+++ grpc-service-sample/src/main/java/com/nmatute/grpcservicesample/GrpcServiceSample.java
@@ -1,14 +1,14 @@
-package com.faxterol.grpcdemo;
+package com.nmatute.grpcservicesample;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
-public class GrpcdemoApplication {
+public class GrpcServiceSample {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GrpcdemoApplication.class, args);
+		SpringApplication.run(GrpcServiceSample.class, args);
 	}
