+++ grpc-service-sample/src/main/java/com/nmatute/grpcservicesample/grpc/interceptor/RequestInterceptor.java
@@ -1,4 +1,4 @@
-package com.faxterol.grpcdemo;
+package com.nmatute.grpcservicesample.grpc.interceptor;
 
 import org.lognet.springboot.grpc.GRpcGlobalInterceptor;
 
@@ -9,7 +9,7 @@ import io.grpc.ServerCallHandler;
 import io.grpc.ServerInterceptor;
 
 @GRpcGlobalInterceptor
-public class GrpcDemoRequestInterceptor implements ServerInterceptor{
+public class RequestInterceptor implements ServerInterceptor{
 
     @Override
     public <ReqT, RespT> Listener<ReqT> interceptCall(ServerCall<ReqT, RespT> call, Metadata headers,
