+++ grpc-service-sample/src/main/java/com/nmatute/grpcservicesample/grpc/stream/MultiplesMensajesStream.java
@@ -1,23 +1,23 @@
-package com.faxterol.grpcdemo;
+package com.nmatute.grpcservicesample.grpc.stream;
 
 import java.util.Random;
 
-import com.faxterol.grpcdemo.interfaces.EnviarMensaje;
-import com.faxterol.grpcdemo.interfaces.RecibirMensaje;
+import com.nmatute.grpcinterface.generated.SendMessage;
+import com.nmatute.grpcinterface.generated.ReceiveMessage;
 
 import io.grpc.stub.StreamObserver;
 
-public class MultiplesMensajesStream implements StreamObserver<EnviarMensaje>{
-    private StreamObserver<RecibirMensaje> responseObserver;
+public class MultiplesMensajesStream implements StreamObserver<SendMessage>{
+    private StreamObserver<ReceiveMessage> responseObserver;
     private Random random = new Random();
 
     //Necesitamos inyectar el stream observer para poder responder al cliente el mensaje
-    public MultiplesMensajesStream(StreamObserver<RecibirMensaje> responseObserver){
+    public MultiplesMensajesStream(StreamObserver<ReceiveMessage> responseObserver){
         this.responseObserver = responseObserver;
     }
 
     @Override
-    public void onNext(EnviarMensaje value) {
+    public void onNext(SendMessage value) {
         System.out.println("MultiplesMensajesStream - Nuevo mensaje recibido: "+value);
         
     }
@@ -31,7 +31,7 @@ public class MultiplesMensajesStream implements StreamObserver<EnviarMensaje>{
     @Override
     public void onCompleted() {
         //Los streams unidireccionales de cliente a servidor, solo requieren una respuesta del servidor.
-        RecibirMensaje respuesta = RecibirMensaje.newBuilder()
+        ReceiveMessage respuesta = ReceiveMessage.newBuilder()
                                     .setFrom(random.nextInt())
                                     .setMessage("Hola mundo desde el servidor: "+random.nextInt())
                                     .build();
