+++ grpc-service-sample/src/main/java/com/nmatute/grpcservicesample/grpc/stream/MultiplesRespuestasStream.java
@@ -1,30 +1,30 @@
-package com.faxterol.grpcdemo;
+package com.nmatute.grpcservicesample.grpc.stream;
 
 import java.util.Random;
 
-import com.faxterol.grpcdemo.interfaces.EnviarMensaje;
-import com.faxterol.grpcdemo.interfaces.RecibirMensaje;
+import com.nmatute.grpcinterface.generated.SendMessage;
+import com.nmatute.grpcinterface.generated.ReceiveMessage;
 
 import io.grpc.stub.StreamObserver;
 
-public class MultiplesRespuestasStream implements StreamObserver<EnviarMensaje>{
-    private StreamObserver<RecibirMensaje> responseObserver;
+public class MultiplesRespuestasStream implements StreamObserver<SendMessage>{
+    private StreamObserver<ReceiveMessage> responseObserver;
     private Random random = new Random();
 
     //Necesitamos inyectar el stream observer para poder responder al cliente el mensaje
-    public MultiplesRespuestasStream(StreamObserver<RecibirMensaje> responseObserver){
+    public MultiplesRespuestasStream(StreamObserver<ReceiveMessage> responseObserver){
         this.responseObserver = responseObserver;
     }
 
 
     @Override
-    public void onNext(EnviarMensaje value) {
+    public void onNext(SendMessage value) {
         System.out.println("MultiplesMensajesStream - Nuevo mensaje recibido: "+value);
 
         //En cuanto recibimos un mensaje, podemos enviar muchos mensajes de regreso al cliente
-        RecibirMensaje mensajeRespuesta = null;
+        ReceiveMessage mensajeRespuesta = null;
         for(int i=0;i<10;i++){
-            mensajeRespuesta = RecibirMensaje.newBuilder()
+            mensajeRespuesta = ReceiveMessage.newBuilder()
                                 .setFrom(random.nextInt())
                                 .setMessage("Hola stream bidireccional desde servidor "+random.nextInt())
                                 .build();
@@ -43,9 +43,9 @@ public class MultiplesRespuestasStream implements StreamObserver<EnviarMensaje>{
 
     @Override
     public void onCompleted() {
-        RecibirMensaje mensajeRespuesta = null;
+        ReceiveMessage mensajeRespuesta = null;
         for(int i=0;i<10;i++){
-            mensajeRespuesta = RecibirMensaje.newBuilder()
+            mensajeRespuesta = ReceiveMessage.newBuilder()
                                 .setFrom(random.nextInt())
                                 .setMessage("Hola stream bidireccional desde servidor al completar "+random.nextInt())
                                 .build();
