+++ grpc-service-sample/src/main/resources/application.yml
@@ -1,0 +1,6 @@
+server:
+  port: 8081
+
+spring:
+  main:
+    banner-mode: off
