--- grpc-service/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.application.name="Spring gRPC"
-grpc.port=6999
