--- grpc-service/src/test/java/com/faxterol/grpcdemo/GrpcdemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.faxterol.grpcdemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GrpcdemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
